added support for WAKE on LAN and expiration to for FING info
the plugin now supports:
send wake on lan pacjkets to ipdevices. it has to be set for each device individually. it "costs" ~ 0.3 secs per device and cycle

added expiration timer to fing up --> down--> expired. the info for down / expired is delayed by xx seconds  to the plugin. In case fing gets new UP info abou the device the timer gets reset.
In a test an iphone 7 and X stayed up for 24 hours w/o inetrruption when WakeOnLan was enabled too.